Remove index signature from CSSProperties
This allows for actually using strongly typed hyphen properties in CSS. In the current version, the comment above it is wrong. This was likely copied from the React types and never fixed to reflect the fact that the index signature was included back: https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/react/index.d.ts#L1577-L1584

The reason for including this signature is not clear to me. Does the project require some hyphen properties to be available? In that case, then Vue should define them explicitly. Otherwise, this makes the types measurably worse for everyone else using Vue's types in their own projects.